+++ docs/CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Added view to reset style_modification_md5 (so template is considered as not modified).
   [sgeulette]
+- Added view to list all templates
+  [sgeulette]
 
 3.0.2 (2017-10-06)
 ------------------